IOS Patcher Update - v1.5.7
I stopped making a new updates for IOS Patcher because i know that People's don't need it and I know that it's 100% completed and it's working like a charm. No more "No internet connection" errors. But it would suck because Fun4TubeGr sent me an translation few weeks ago so you know, i had to add it.